Update Disk model to use Name value object
This commit changes the Disk model to use the new Name value object instead of a string for the Name property.  This provides better encapsulation and allows for validation rules to be centralized within the Name value object. It also updates various tests, configurations, handlers and repository methods to reflect this change. A local database source is added for local testing.