Refactor and add new tests
Numerous changes have been made including adding new integration tests for DiskRepository and unit tests for Reviews, a reference to Tests.Domain in Domain.Tests.csproj, and changing the database configuration to use in-memory database. Also, added validation and logic to Review and Disk domain classes and configured pseudonym to be mandatory in AuthorConfiguration. Several other minor changes and fixes are also included.