Update database model and switch to SQL Server
The database model `ApplicationDbContextModelSnapshot.cs` has been updated with modifications to the "Pseudonyme" property and the addition of the "Review" entity. The migration file for renaming the "Pseudonyme" property to "Pseudonym" has been added. Also, the application's database configuration has been switched from an in-memory database to SQL Server in `DependencyInjection.cs`.