Update various parts of Domain, Data, Service and API tests
Code changes encompass the removal of unused files and adjustment of some existing ones. In the domain, the AggregateRoot and Entity classes were simplified, the Title file was removed, and some test fixtures were created. In the data layer, the EntitySaveChangesInterceptor and ApplicationDbContextInitializer were modified to integrate with the domain-driven design (DDD) better.