Add SqlConnection handling and refactor services
The SqlConnection object is now created and managed within the ServiceFixtureBase class. This involves opening the connection when needed and closing it during disposal. Additionally, some services were refactored, particularly the usage of GetRequiredService in test fixtures and the removal of service additions in the DependencyInjection file. These have been moved to the Program file for improved organization.